Start supporting expression in where clauses
Add grammar rules for full expressions (with logical operators and
comparison operators). Use PrecClimber to evaluate expressions in the
context of a row. Change where_clause to take an expression instead of a
hardcoded equality. Add SelectQueryResult.filter_expression to filter
rows by whether an expression is true or not.